more

+++ packages/protocol/tasks/deploy_L1.ts
@@ -197,7 +197,7 @@ export async function deployContracts(hre: any) {
             `${chainId}.${Buffer.from(
                 ethers.utils.arrayify(
                     ethers.utils.solidityPack(
-                        ["string", "uint256", "string", "uint8"],
+                        ["string", "uint256", "string", "uint16"],
                         ["plonk_verifier_", 0, "_", 0]
                     )
                 )
